requirement.txt
flask==3.0.0
pillow==10.1.0
torch==2.1.1
torchvision==0.16.1
open-clip-torch==2.23.0
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
